--- src/App.module.css
@@ -1,6 +1,0 @@
-.flex {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-}
