+++ src/components/MenuHeader/index.js
@@ -1,0 +1,17 @@
+import {useState} from 'react'
+import Menu from "../Menu";
+import Navbar from "../Navbar";
+
+const MenuHeader = () => {
+    const [isMenu, setMenu] = useState(false)
+    const handleMenu = (menuState) => {
+        setMenu(menuState)
+    }
+    return (
+        <>
+            <Menu menuState={isMenu}/>
+            <Navbar onMenu={handleMenu} menuState={isMenu}/>
+        </>
+    )
+}
+export default MenuHeader
